fix(): resolved the issue of bad annotations
There was the issue that annotations would not get properly annotated.
This has been resolved with this fix by using the background script to
act as a buffer store. The annotation will be stored inside the session
storage of the extension. The moment react finishes hydrating it will
check the storage for the annotation.